Simplify append procedure
Previously, I had implemented append as a foldr on a lambda
that would take first and init as arguments and cons the two.
Now I realized its pointless to write a lambda wrapper over the
cons procedure. So now I call foldr on cons with list-b as init
and list-a as the list to be traversed. This makes the append
implementation simple and concise, as all solutions should be.