Add procedures to remove element from list
This commit adds a couple of primitives to remove an element
from given list. The first procedure is my own implementation of
the Scheme's standard remove procedure. It removes all occurrences
of an element in the list. The second is a bit different in that
it removes only the first occurence of an element. The rest are
unchecked. The second procedure was written to emulate Python's
list.remove(ele) function.