Start implementing frame as a list of bindings
Previously, frames were represented as a pair of lists with the
car holding the list of variables and cdr holding the list of
values. Now I'm trying to represent a frame as a list of
bindings where each binding is a name-value pair. This
representation is similar to how we did it in COMPSCI220 and also
in Lispy.