Add lookup variable for new environment
This implementation needs a lot of bug fixing to make it work
correctly. The general structure of variable lookup seems to be
right. The bugs arise from accessing the wrong parts of the
bindings in a frame. For fixing, I should keep in mind that the
initial environment is an extension of empty environment.